+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Configurable Kafka trace correlation (enabled/disabled) and correlation header format (string, binary, both).
 - [AWS Lambda] Add DynamoDB streams trigger.
 - [W3C Trace Context] Reject traceparent header when either trace ID or parent span ID are all zeroes.
+- [Bull] Fixed the error `TypeError: Cannot read property 'opts' of null` in repeatable jobs.
 
 ## 1.136.0
 - An issue event is sent if the application uses an EOL (end of life) version of Node.js. Applicable only for non serverless environments.